// src/app/enums/platforms.enum.ts
export enum PlatformsEnum {
  GitHub ='GITHUB',
  FrontendMentor = 'FRONTEND_MENTOR',
  Twitter = 'TWITTER',
  LinkedIn = 'LINKEDIN',
  YouTube = 'YOU_TUBE',
  Facebook = 'FACEBOOK',
  Twitch = 'TWITCH',
  DevTO = 'DEV_TO',
  Codewars = 'CODEWARS',
  Codepen = 'CODEPEN',
  FreeCodeCamp = 'FREE_CODE_CAMP',
  GitLab = 'GITLAB',
  Hashnode = 'HASHNODE',
  StackOverflow = 'STACK_OVERFLOW'
}